fix(tests): a test should not cause side-effects in other tests (#23)
If a fixture object was modified in one test, the modification would be visible
to other tests using the same fixture object.

I encountered this problem whilst testing an upcoming PR. This change makes
every usage of a fixture a new object. I'm open to suggestions on alternative
solutions, such as specifying that fixture modification is a test failure.